requirements.txt
playwright
google-generativeai
beautifulsoup4
python-dotenv
Pillow